==> code/06_compliance_analysis.ipynb <==
# Analysing mHealth Apps compliance with privacy laws or regulations

import pandas as pd

merged_df = pd.read_csv('../data/merged_df.csv')
all_answers = pd.read_csv('../data/RQ_answers.csv')

mhealth_apps_final = pd.merge(merged_df, all_answers, on='app_id', how='left')
mhealth_apps_final['a12'] = mhealth_apps_final['a12'].fillna('Omitted') # apps lacking privacy policy

mhealth_apps_final.to_csv('../data/mhealth_apps_final.csv', index=False)

# Q1: Does the privacy policy specify the types of data collected?

app_data_collected = pd.read_csv('../data/app_data_collected.csv')

print('Number of Apps mentioning collecting data in Data Safety:', len(app_data_collected['app_id'].unique()))

print('\nNumber of Apps mentioning collecting data in Privacy Policy:')
a1_data_collected = mhealth_apps_final['a1'].value_counts()
print(a1_data_collected)

print('\nVariances:', len(app_data_collected['app_id'].unique()) - len(mhealth_apps_final[mhealth_apps_final['a1'] == 'Specified']))

# Q2: Does the policy explain data collection practices for all, partial, or none of the collected data?

print('\nNumber of Apps explaining data collection:')
a2_collection_purpose = mhealth_apps_final['a2'].value_counts()
a2_collection_purpose

# Q3: Does the privacy policy specify the types of data shared with third parties?

app_data_shared = pd.read_csv('../data/app_data_shared.csv')
print('Number of Apps mentioning sharing data in Data Safety:', len(app_data_shared['app_id'].unique()))

print('\nNumber of Apps mentioning sharing data in Privacy Policy:')
a3_data_shared = mhealth_apps_final['a3'].value_counts()
print(a3_data_shared)

print('\nVariances:', len(app_data_shared['app_id'].unique()) - len(mhealth_apps_final[mhealth_apps_final['a3'] == 'Specified']))

# Q4: Does the policy explain data sharing practices for all, partial, or none of the shared data?

print('\nNumber of Apps explaining data sharing:')
a4_sharing_purpose = mhealth_apps_final['a4'].value_counts()
a4_sharing_purpose

# Q5: Does the policy explain who is receiving the shared data?

print('\nNumber of Apps explaining receipients of sharing data:')
a5_sharing_recipients = mhealth_apps_final['a5'].value_counts()
print(a5_sharing_recipients)

print('\nNumber of Apps sharing data but failed to explain recipients of shared data:')
print(len(mhealth_apps_final[(mhealth_apps_final['a3'] == 'Specified') & (mhealth_apps_final['a5'] == 'No')]))

# Q6: Does the policy explain users’ rights to access, correct, delete, or opt out of marketing communications?

a6_user_rights = mhealth_apps_final['a6'].value_counts()
a6_user_rights

# Q7: Does the policy mention if the data is encrypted in transit?

app_security_practices = pd.read_csv('../data/app_security_practices.csv')

print('Number of Apps mentioning data encrypted in Data Safety:', len(app_security_practices[app_security_practices['security_practices'] == 'Data is encrypted in transit']))

a7_encrypted = mhealth_apps_final['a7'].value_counts()
a7_encrypted

# Q8: Does the policy outline security measures against unauthorized access or breaches?

a8_security_measures = mhealth_apps_final['a8'].value_counts()
a8_security_measures

# Q9: Is contact information provided for questions or complaints? 

print('\nNumber of Apps provide information for questions or complaints:')
a9_contract_info = mhealth_apps_final['a9'].value_counts()
a9_contract_info

# Q10: Does the policy address children’s data and parental consent? 

a10_children = mhealth_apps_final['a10'].value_counts()
a10_children

# Q11: Does the policy explain the compliance with privacy laws?

import pandas as pd

usa = ['us']

# source: https://www.gdpradvisor.co.uk/gdpr-countries
GDPR_countries = ['Austria', 'Belgium', 'Bulgaria', 'Croatia', 'Cyprus', 'Czechia', 'Denmark', 'Estonia', 'Finland', 'France', 
                  'Germany', 'Greece', 'Hungary', 'Ireland', 'Italy', 'Latvia', 'Lithuania', 'Luxembourg', 'Malta', 'Netherlands', 
                  'Poland', 'Portugal', 'Romania', 'Slovakia', 'Slovenia', 'Spain', 'Sweden', 'United Kingdom']

country_list = pd.read_csv("../data/country_list.csv")

europe = []

for country in GDPR_countries:
    matching_countries = country_list[country_list['country_name'].str.contains(country, na=False)]
    europe.extend(matching_countries['country_code'].tolist())

print(len(europe)-1)
print(europe)

def filter_data(df):

    def country_filter(countries, country_set):
        return any(country.lower() in country_set for country in countries)

    countries_list = df['countries'].str.split(', ').explode()
    country_counts = countries_list.value_counts()
    top_5_countries = country_counts.head(5)
    print("Top 5 countries:")
    print(top_5_countries)
    
    usa_europe = set(usa + europe)
    rest_of_world = set(countries_list) - usa_europe
    
    df = df.copy()
    df['country_split'] = df['countries'].str.split(', ')

    # Apply filters
    us_filtered_df = df[df['country_split'].apply(lambda x: country_filter(x, usa))]
    europe_filtered_df = df[df['country_split'].apply(lambda x: country_filter(x, europe))]
    row_filtered_df = df[df['country_split'].apply(lambda x: country_filter(x, rest_of_world))]

    print('Number of apps available in USA:', len(us_filtered_df))
    print('Number of apps available in Europe:', len(europe_filtered_df))
    print('Number of apps available in RoW:', len(row_filtered_df))
    
    return us_filtered_df, europe_filtered_df, row_filtered_df

# Inconsistency Assessment

# Apps claim to have a privacy policy but fail to provide one
total_inconsist_apps = mhealth_apps_final[
    (mhealth_apps_final['is_privacy_policy'] != mhealth_apps_final['privacy_policy_file']) |
    ((mhealth_apps_final['a1'] == 'Not mentioned') & mhealth_apps_final['data_collected'].notnull()) |
    ((mhealth_apps_final['a3'] == 'Not mentioned') & mhealth_apps_final['data_shared'].notnull()) |
    ((mhealth_apps_final['security_practices'].fillna('').str.contains('Data is encrypted')) & (mhealth_apps_final['a7'] != 'Encrypted')) |
    ((mhealth_apps_final['data_shared'].isnull() | mhealth_apps_final['data_collected'].isnull()) & 
     (mhealth_apps_final['PII'].notnull() | 
      mhealth_apps_final['PHI'].notnull() | 
      mhealth_apps_final['PII&PHI'].notnull()))
]

print(f'Apps contain inconsistency: {len(total_inconsist_apps)} ({len(total_inconsist_apps)/len(mhealth_apps_final)*100:.1f}%)')

import numpy as np
import pandas as pd

relevant_data = set(
    item.strip().lower()
    for col in ['data_shared', 'data_collected', 'a1_note', 'a3_note']
    for sublist in mhealth_apps_final[col].dropna().astype(str).str.split(',')
    for item in sublist
)

# Check for inconsistencies with case-insensitive comparison
inconsistent_rows = mhealth_apps_final[
    mhealth_apps_final['PII'].apply(
        lambda pii: pd.notna(pii) and any(
            pii_item.strip().lower() not in relevant_data 
            for pii_item in pii.split(',')
        )
    )
]

inconsistent_apps = inconsistent_rows
print('Number of apps collecting/sharing PII that is not mentioned in their Data Safety section or privacy policies')
print('PII Inconsistency:', len(inconsistent_apps))

using_other_ids = inconsistent_apps[
    (inconsistent_apps['data_shared'].str.contains("other IDs", case=False, na=False) |
     inconsistent_apps['data_collected'].str.contains("other IDs", case=False, na=False) |
     inconsistent_apps['a1_note'].str.contains("other IDs", case=False, na=False) |
     inconsistent_apps['a3_note'].str.contains("other IDs", case=False, na=False))
    &
    (inconsistent_apps['PII'].str.contains("aaid|aid|imsi|mac", case=False, na=False))
]
print(len(using_other_ids))

inconsistent_apps = inconsistent_apps.merge(using_other_ids, how='outer', indicator=True).query('_merge == "left_only"').drop(columns=['_merge'])
print(len(inconsistent_apps))

cdc_apps = inconsistent_apps[inconsistent_apps['app_id'].str.contains('cdc', na=False)]
print('Number of CDC apps:', len(cdc_apps))

category_counts = inconsistent_apps.groupby('categories').size().reset_index(name='total_count')
category_counts = category_counts.sort_values(by='total_count', ascending=False).reset_index(drop=True)

category_counts

us_filtered_df, europe_filtered_df, row_filtered_df = filter_data(inconsistent_apps)
print('Average downloads in USA:', us_filtered_df['downloads'].mean())
print('Average downloads in Europe:', europe_filtered_df['downloads'].mean())
print('Average downloads in RoW:', row_filtered_df['downloads'].mean())


# Noncompliance Assessment

import re

def extract_unique_item(dataframe):
    unique_items = set()
    for lists in dataframe.dropna():
        item_list = lists.split(',')
        for items in item_list:
            items = items.split(',')
        unique_items.update(item.strip() for item in items if item)
    return unique_items

unique_laws = extract_unique_item(mhealth_apps_final['a11'])
print('Total unique laws:', len(unique_laws))
unique_laws


no_compliance = mhealth_apps_final[mhealth_apps_final['a11'] == 'Not mentioned']
print("Apps do not mentioned laws or regulations:", len(no_compliance))
no_compliance_count = no_compliance.groupby('categories').size()
no_compliance_count

us_filtered_df, europe_filtered_df, row_filtered_df = filter_data(no_compliance)
filtered_no_compliance = pd.concat([us_filtered_df, europe_filtered_df], ignore_index=True)
print('Average downloads in USA:', us_filtered_df['downloads'].mean())
print('Average downloads in Europe:', europe_filtered_df['downloads'].mean())
print('Average downloads in RoW:', row_filtered_df['downloads'].mean())

noncompliance_apps = filtered_no_compliance[
    filtered_no_compliance[['data_shared', 'data_collected', 'PII', 'PHI', 'PII&PHI']].isnull().all(axis=1)
]

print("Apps with no compliance:\n")
apps_sharing = filtered_no_compliance[filtered_no_compliance['data_shared'].notnull()]
print('Sharing data:', len(apps_sharing))

apps_collecting = filtered_no_compliance[filtered_no_compliance['data_collected'].notnull()]
print('Collecting data:', len(apps_collecting))

apps_pii = filtered_no_compliance[filtered_no_compliance['PII'].notnull()]
print('Collecting/sharing PII:', len(apps_pii))

apps_phi = filtered_no_compliance[filtered_no_compliance['PHI'].notnull()]
print('Collecting/sharing PHI:', len(apps_phi))

apps_pii_phi = filtered_no_compliance[filtered_no_compliance['PII&PHI'].notnull()]
print('Collecting/sharing PII&PHI:', len(apps_pii_phi))

law_mentioned_apps = mhealth_apps_final[mhealth_apps_final['a11'] != 'Not mentioned']
print(f'Number of apps mentioning laws in their privacy policies: {len(law_mentioned_apps)}\n')

us_filtered_df, europe_filtered_df, row_filtered_df = filter_data(law_mentioned_apps)
law_mentioned_apps = pd.concat([us_filtered_df, europe_filtered_df], ignore_index=True)
print('Average downloads in USA:', us_filtered_df['downloads'].mean())
print('Average downloads in Europe:', europe_filtered_df['downloads'].mean())
print('Average downloads in RoW:', row_filtered_df['downloads'].mean())

unique_laws_us = extract_unique_item(us_filtered_df['a11'])
print('Total unique laws:', len(unique_laws_us))

no_HIPAA = us_filtered_df[~us_filtered_df['a11'].str.contains('HIPAA', case=False, na=False)]
print(f'Apps that do not mention HIPAA: {len(no_HIPAA)} ({len(no_HIPAA) / len(us_filtered_df) * 100:.2f}%)')


print("Apps accessible in USA:\n")

apps_sharing = no_HIPAA[no_HIPAA['data_shared'].notnull()]
print('Sharing data:', len(apps_sharing))

apps_collecting = no_HIPAA[no_HIPAA['data_collected'].notnull()]
print('Collecting data:', len(apps_collecting))

apps_pii = no_HIPAA[no_HIPAA['PII'].notnull()]
print('Collecting/sharing PII:', len(apps_pii))

apps_phi = no_HIPAA[no_HIPAA['PHI'].notnull()]
print('Collecting/sharing PHI:', len(apps_phi))

apps_pii_phi = no_HIPAA[no_HIPAA['PII&PHI'].notnull()]
print('Collecting/sharing PII&PHI:', len(apps_pii_phi))

unique_laws_europe = extract_unique_item(europe_filtered_df['a11'])
print('Total unique laws:', len(unique_laws_europe))
no_GDPR = europe_filtered_df[~europe_filtered_df['a11'].str.contains('GDPR', case=False, na=False)]
print(f'Apps that do not mention GDPR: {len(no_GDPR)} ({len(no_GDPR) / len(europe_filtered_df) * 100:.2f}%)')


print("Apps accessible in Europe:\n")

apps_sharing = no_GDPR[no_GDPR['data_shared'].notnull()]
print('Sharing data:', len(apps_sharing))

apps_collecting = no_GDPR[no_GDPR['data_collected'].notnull()]
print('Collecting data:', len(apps_collecting))

apps_pii = no_GDPR[no_GDPR['PII'].notnull()]
print('Collecting/sharing PII:', len(apps_pii))

apps_phi = no_GDPR[no_GDPR['PHI'].notnull()]
print('Collecting/sharing PHI:', len(apps_phi))

apps_pii_phi = no_GDPR[no_GDPR['PII&PHI'].notnull()]
print('Collecting/sharing PII&PHI:', len(apps_pii_phi))

# COPPA Compliance Assessment

irrelevant_data = ['Teen', 'Mature 17+', 'Rated 12+', 'Rated for 12+', 'Rated for 18+']
apps_coppa = mhealth_apps_final[
    mhealth_apps_final['content_rating'].notna() & 
    ~mhealth_apps_final['content_rating'].isin(irrelevant_data)
]
print('Number of apps under COPPA:', len(apps_coppa))

a10_children = apps_coppa['a10'].value_counts()

print('Average downloads of Detailed:', apps_coppa[apps_coppa['a10'] == 'Detailed']['downloads'].mean())
print('Average downloads of Brief:', apps_coppa[apps_coppa['a10'] == 'Brief']['downloads'].mean())
print('Average downloads of Not mentioned:', apps_coppa[apps_coppa['a10'] == 'Not mentioned']['downloads'].mean())

a10_children

no_coppa_apps = apps_coppa[
    (apps_coppa['a10'] == 'Not mentioned') & 
    (
        (apps_coppa['data_shared'].notnull()) | 
        (apps_coppa['data_collected'].notnull()) | 
        (apps_coppa['a1'] == 'Yes') | 
        (apps_coppa['a3'] == 'Yes') | 
        (apps_coppa['PII'].notnull()) | 
        (apps_coppa['PHI'].notnull()) | 
        (apps_coppa['PII&PHI'].notnull())
    )
]

print('Number of apps sharing/collecting data but failed to address COPPA:', len(no_coppa_apps))
us_filtered_df, europe_filtered_df, row_filtered_df= filter_data(no_coppa_apps)
print('\nAverage downloads in USA:', us_filtered_df['downloads'].mean())
print('Average downloads in Europe:', europe_filtered_df['downloads'].mean())
print('\nAverage downloads in RoW:', row_filtered_df['downloads'].mean())


no_coppa_apps_category = no_coppa_apps['categories'].value_counts()
no_coppa_apps_category

# Q12: How would you rate the policy disclosure?

print('\nRate of the policy disclosure:')
a12_disclosure = mhealth_apps_final['a12'].value_counts()
a12_disclosure
