==> code/07_case_study_&_impact_factors.ipynb <==
# Case Study & Impact Factors Analysis

import pandas as pd

def process_category_data(dataframe, top_n=5):
    # Step 1: Calculate raw category counts
    category_counts_raw = dataframe.groupby('categories').size().reset_index(name='total_count')
    category_counts_raw = category_counts_raw.sort_values(by='total_count', ascending=False).reset_index(drop=True)

    # Step 2: Define aggregation logic
    agg_logic = {
        'RQ1': ('a1', lambda x: round(
            (((x == 'Specified') & dataframe['data_collected'].notnull()).sum() + ((x == 'Not mentioned') & dataframe['data_collected'].isnull()).sum()) / len(x), 2)),
        'RQ2': ('a3', lambda x: round(
            (((x == 'Specified') & dataframe['data_shared'].notnull()).sum() + ((x == 'Not mentioned') & dataframe['data_shared'].isnull()).sum()) / len(x), 2)),
        'RQ3': ('a6', lambda x: round(
            ((x != 'Not at all') & dataframe['security_practices'].str.contains('data be deleted')).sum() / len(x), 2)),
        'RQ4': ('a7', lambda x: round(
            (((x == 'Encrypted') & dataframe['security_practices'].str.contains('is encrypted')).sum() + 
             ((x != 'Encrypted') & dataframe['data_shared'].isnull() & dataframe['data_collected'].isnull() 
              & (dataframe['a1'] == 'No') & (dataframe['a3'] == 'No')).sum()) / len(x), 2)),
        'RQ5': ('a8', lambda x: round(((x != 'Not mentioned').sum() / len(x)), 2)),
        'RQ6': ('a9', lambda x: round(((x == 'Yes').sum() / len(x)), 2)),
        'RQ7': ('a10', lambda x: round(((x != 'Not mentioned').sum() / len(x)), 2)),
        'RQ8': ('a12', lambda x: round(((x != 'Omitted').sum() / len(x)), 2)),
        'RQ9': ('a11', lambda x: round(((x != 'Not mentioned').sum() / len(x)), 2))
    }
    
    # Step 3: Aggregate initial statistics
    RQ_counts = dataframe.groupby('categories').agg(**agg_logic).reset_index()
    RQ_counts = RQ_counts.merge(category_counts_raw, on='categories')
    RQ_counts = RQ_counts.sort_values(by='total_count', ascending=False).reset_index(drop=True)

    # Step 4: Identify top categories
    top_categories = RQ_counts['categories'].head(top_n)
    
    # Step 5: Categorize remaining into "Others"
    dataframe.loc[:, 'categories'] = dataframe['categories'].apply(
        lambda x: x if x in top_categories.values else 'Others'
    )
    
    # Step 6: Recalculate statistics with "Others" category
    RQ_counts = dataframe.groupby('categories').agg(
        **agg_logic,
        total_count=('categories', 'size')
    ).reset_index()
    RQ_counts = RQ_counts.sort_values(by='total_count', ascending=False).reset_index(drop=True)
    
    return RQ_counts

mhealth_apps_final = pd.read_csv('../data/mhealth_apps_final.csv')
mhealth_processed = process_category_data(mhealth_apps_final)
mhealth_processed

# Case Study - CDC Apps

cdc_apps = mhealth_apps_final[mhealth_apps_final['app_id'].str.contains('cdc', na=False)]
print('Number of CDC apps:', len(cdc_apps))
print('Average downloads:', cdc_apps['downloads'].mean())

import os

privacy_policies_dir = '../data/privacy_policies'

cdc_policies = []

for app_id in cdc_apps['app_id']:
    for filename in os.listdir(privacy_policies_dir):
        if app_id in filename:
            cdc_policies.append(filename)

print('Found privacy policies:', len(cdc_policies))


cdc_apps_processed = process_category_data(cdc_apps)
cdc_apps_processed

# Impact Factors Analysis

## Paid Apps

paid_apps = mhealth_apps_final[mhealth_apps_final['free'] == False]
print('Total paid apps:', len(paid_apps))

import os

traffic_data_dir = '../data/network_traffic'

paid_app_traffic = []

for app_id in paid_apps['app_id']:
    for filename in os.listdir(traffic_data_dir):
        if app_id in filename:
            app_data = paid_apps[paid_apps['app_id'] == app_id]
            if app_data['PII'].isna().all() and \
               app_data['PHI'].isna().all() and \
               app_data['PII&PHI'].isna().all() and \
               app_data['Connectivity Issues'].eq('No Content').all():
                continue
            paid_app_traffic.append(filename)

print('Total paid apps with traffic data:', len(paid_app_traffic))


paid_apps_processed = process_category_data(paid_apps)
paid_apps_processed

## Free Apps

free_apps = mhealth_apps_final[mhealth_apps_final['free'] == True]
print('Total free apps:', len(free_apps))

free_apps_processed = process_category_data(free_apps)
free_apps_processed

## Top-grossing Apps

top_grossing_apps = mhealth_apps_final[mhealth_apps_final['top_grossing'] == 'Yes']
print('Total top-grossing apps:', len(top_grossing_apps))

top_grossing_apps_processed = process_category_data(top_grossing_apps)
top_grossing_apps_processed

## Non-top-grossing Apps

non_top_grossing_apps = mhealth_apps_final[mhealth_apps_final['top_grossing'] != 'Yes']
print('Total non-top-grossing apps:', len(non_top_grossing_apps))

non_top_grossing_apps_processed = process_category_data(non_top_grossing_apps)
non_top_grossing_apps_processed

## Geolocations: US, EU, RoW

import pandas as pd

usa = ['us']

# source: https://www.gdpradvisor.co.uk/gdpr-countries
GDPR_countries = ['Austria', 'Belgium', 'Bulgaria', 'Croatia', 'Cyprus', 'Czechia', 'Denmark', 'Estonia', 'Finland', 'France', 
                  'Germany', 'Greece', 'Hungary', 'Ireland', 'Italy', 'Latvia', 'Lithuania', 'Luxembourg', 'Malta', 'Netherlands', 
                  'Poland', 'Portugal', 'Romania', 'Slovakia', 'Slovenia', 'Spain', 'Sweden', 'United Kingdom']

country_list = pd.read_csv("../data/country_list.csv")

europe = []

for country in GDPR_countries:
    matching_countries = country_list[country_list['country_name'].str.contains(country, na=False)]
    europe.extend(matching_countries['country_code'].tolist())

print(len(europe)-1)
print(europe)

def filter_data(df):

    def country_filter(countries, country_set):
        return any(country.lower() in country_set for country in countries)

    countries_list = df['countries'].str.split(', ').explode()
    usa_europe = set(usa + europe)
    rest_of_world = set(countries_list) - usa_europe
    
    df['country_split'] = df['countries'].str.split(', ')
    us_filtered_df = df[df['country_split'].apply(lambda x: country_filter(x, usa))]
    europe_filtered_df = df[df['country_split'].apply(lambda x: country_filter(x, europe))]
    row_filtered_df = df[df['country_split'].apply(lambda x: country_filter(x, rest_of_world))]

    print('Number of apps available in USA:', len(us_filtered_df))
    print('Number of apps available in Europe:', len(europe_filtered_df))
    print('Number of apps available in RoW:', len(row_filtered_df))
    
    return us_filtered_df, europe_filtered_df, row_filtered_df


us_filtered_df, europe_filtered_df, row_filtered_df = filter_data(mhealth_apps_final)

### 1. USA

us_apps_processed = process_category_data(us_filtered_df, 8)
us_apps_processed

### 2. Europe

europe_apps_processed = process_category_data(europe_filtered_df, 8)
europe_apps_processed

### 3. RoW (Rest of the World)

row_apps_processed = process_category_data(row_filtered_df, 8)
row_apps_processed
